requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
MetaTrader5==5.0.45
pandas==2.1.4
numpy==1.24.3
python-multipart==0.0.6

# 部署脚本依赖
requests==2.31.0
cryptography==41.0.7
paramiko==3.3.1
pytest==7.4.3
